+++ core/module.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var async = require('async');
 var fs = require('fs');
 var modDriver = require('./driver');