+++ scripts/daemon/start.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var run = require('child_process').spawn;
 var gn = require('gracenode');
 var logger = gn.log.create('daemon-start');